refactor(fonts): match cli structure
