+++ packages/astro/src/assets/fonts/README.md
@@ -9,10 +9,10 @@ Here is an overview of the architecture of the fonts in Astro:
   - For each family, it resolves fonts data and normalizes them
   - For each family, optimized fallbacks (and related CSS) are generated if applicable
   - It returns the data
-- [`/logic`](./logic/) contains the sub steps of `orchestrate()` so they can be easily tested
+- [`/core`](./core/) contains the sub steps of `orchestrate()` so they can be easily tested
 - The logic uses [inversion of control](https://en.wikipedia.org/wiki/Inversion_of_control) to make it easily testable and swappable
   - [`definitions.ts`](./definitions.ts) defines dependencies
-  - Those dependencies are implemented in [`/implementations`](./implementations/)
+  - Those dependencies are implemented in [`/infra`](./infra/)
 - [`fontsPlugin()`](./vite-plugin-fonts.ts) calls `orchestrate()` and using its result, setups anything required so that fonts can
   - Be exposed to users (virtual module)
   - Be used in dev (middleware)
