+++ packages/astro/src/assets/fonts/infra/css-renderer.ts
