+++ packages/astro/src/assets/fonts/infra/error-handler.ts
