+++ packages/astro/src/assets/fonts/infra/font-file-reader.ts
