+++ packages/astro/src/assets/fonts/infra/hasher.ts
