+++ packages/astro/src/assets/fonts/infra/remote-font-provider-mod-resolver.ts
