+++ packages/astro/src/assets/fonts/infra/url-proxy-content-resolver.ts
