+++ packages/astro/src/assets/fonts/infra/url-proxy-hash-resolver.ts
