+++ packages/astro/src/assets/fonts/infra/url-proxy.ts
