+++ packages/astro/test/units/assets/fonts/core.test.js
@@ -1,13 +1,13 @@
 // @ts-check
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import { createFontTypeExtractor } from '../../../../dist/assets/fonts/implementations/font-type-extractor.js';
-import { createSystemFallbacksProvider } from '../../../../dist/assets/fonts/implementations/system-fallbacks-provider.js';
-import { dedupeFontFaces } from '../../../../dist/assets/fonts/logic/dedupe-font-faces.js';
-import { extractUnifontProviders } from '../../../../dist/assets/fonts/logic/extract-unifont-providers.js';
-import { normalizeRemoteFontFaces } from '../../../../dist/assets/fonts/logic/normalize-remote-font-faces.js';
-import { optimizeFallbacks } from '../../../../dist/assets/fonts/logic/optimize-fallbacks.js';
-import { resolveFamily } from '../../../../dist/assets/fonts/logic/resolve-families.js';
+import { dedupeFontFaces } from '../../../../dist/assets/fonts/core/dedupe-font-faces.js';
+import { extractUnifontProviders } from '../../../../dist/assets/fonts/core/extract-unifont-providers.js';
+import { normalizeRemoteFontFaces } from '../../../../dist/assets/fonts/core/normalize-remote-font-faces.js';
+import { optimizeFallbacks } from '../../../../dist/assets/fonts/core/optimize-fallbacks.js';
+import { resolveFamily } from '../../../../dist/assets/fonts/core/resolve-families.js';
+import { createFontTypeExtractor } from '../../../../dist/assets/fonts/infra/font-type-extractor.js';
+import { createSystemFallbacksProvider } from '../../../../dist/assets/fonts/infra/system-fallbacks-provider.js';
 import {
 	createSpyUrlProxy,
 	fakeFontMetricsResolver,
@@ -15,7 +15,7 @@ import {
 	simpleErrorHandler,
 } from './utils.js';
 
-describe('fonts logic', () => {
+describe('fonts core', () => {
 	describe('resolveFamily()', () => {
 		it('removes quotes correctly', async () => {
 			const hasher = { ...fakeHasher, hashObject: () => 'xxx' };
