+++ packages/astro/test/units/assets/fonts/orchestrate.test.js
@@ -5,19 +5,19 @@ import { fileURLToPath } from 'node:url';
 import { defineFontProvider } from 'unifont';
 import { joinPaths } from '../../../../../internal-helpers/dist/path.js';
 import { DEFAULTS } from '../../../../dist/assets/fonts/constants.js';
-import { createMinifiableCssRenderer } from '../../../../dist/assets/fonts/implementations/css-renderer.js';
-import { createDataCollector } from '../../../../dist/assets/fonts/implementations/data-collector.js';
-import { createFontaceFontFileReader } from '../../../../dist/assets/fonts/implementations/font-file-reader.js';
-import { createFontTypeExtractor } from '../../../../dist/assets/fonts/implementations/font-type-extractor.js';
-import { createLevenshteinStringMatcher } from '../../../../dist/assets/fonts/implementations/levenshtein-string-matcher.js';
-import { createRequireLocalProviderUrlResolver } from '../../../../dist/assets/fonts/implementations/local-provider-url-resolver.js';
-import { createBuildRemoteFontProviderModResolver } from '../../../../dist/assets/fonts/implementations/remote-font-provider-mod-resolver.js';
-import { createRemoteFontProviderResolver } from '../../../../dist/assets/fonts/implementations/remote-font-provider-resolver.js';
-import { createSystemFallbacksProvider } from '../../../../dist/assets/fonts/implementations/system-fallbacks-provider.js';
-import { createUrlProxy } from '../../../../dist/assets/fonts/implementations/url-proxy.js';
-import { createRemoteUrlProxyContentResolver } from '../../../../dist/assets/fonts/implementations/url-proxy-content-resolver.js';
-import { createBuildUrlProxyHashResolver } from '../../../../dist/assets/fonts/implementations/url-proxy-hash-resolver.js';
-import { createDevUrlResolver } from '../../../../dist/assets/fonts/implementations/url-resolver.js';
+import { createMinifiableCssRenderer } from '../../../../dist/assets/fonts/infra/css-renderer.js';
+import { createDataCollector } from '../../../../dist/assets/fonts/infra/data-collector.js';
+import { createFontaceFontFileReader } from '../../../../dist/assets/fonts/infra/font-file-reader.js';
+import { createFontTypeExtractor } from '../../../../dist/assets/fonts/infra/font-type-extractor.js';
+import { createLevenshteinStringMatcher } from '../../../../dist/assets/fonts/infra/levenshtein-string-matcher.js';
+import { createRequireLocalProviderUrlResolver } from '../../../../dist/assets/fonts/infra/local-provider-url-resolver.js';
+import { createBuildRemoteFontProviderModResolver } from '../../../../dist/assets/fonts/infra/remote-font-provider-mod-resolver.js';
+import { createRemoteFontProviderResolver } from '../../../../dist/assets/fonts/infra/remote-font-provider-resolver.js';
+import { createSystemFallbacksProvider } from '../../../../dist/assets/fonts/infra/system-fallbacks-provider.js';
+import { createUrlProxy } from '../../../../dist/assets/fonts/infra/url-proxy.js';
+import { createRemoteUrlProxyContentResolver } from '../../../../dist/assets/fonts/infra/url-proxy-content-resolver.js';
+import { createBuildUrlProxyHashResolver } from '../../../../dist/assets/fonts/infra/url-proxy-hash-resolver.js';
+import { createDevUrlResolver } from '../../../../dist/assets/fonts/infra/url-resolver.js';
 import { orchestrate } from '../../../../dist/assets/fonts/orchestrate.js';
 import { defineAstroFontProvider } from '../../../../dist/assets/fonts/providers/index.js';
 import { createSpyLogger, defaultLogger } from '../../test-utils.js';
